Revert "Add spark.pipeline option."
This reverts commit d3452916308fa6ac290e1f410bb425b180cc0f57.

This commit is being reverted because it doesn't work properly for
two reasons:
(1) There's still pipelining in the Hadoop read path because the Hadoop
code pipelines decompression with reading input data.
(2) This results in incorrect data, which is looks like is because Hadoop
reuses the key/value objects, so to make this work, we'll need to explicitly
copy the data rather than just unrolling the iterator.

Conflicts:
	core/src/main/scala/org/apache/spark/rdd/HadoopRDD.scala