Removed Netty shuffle implementation.
This commit removes the netty shuffle implementation (the default
shuffle implementation uses Java's NIO library).

According to Patrick Wendell, this never worked well anyway (there
were a bunch of corner cases it didn't support, and in the more
recent Spark branch this code was completely rewritten:
https://github.com/apache/spark/pull/2753), and having
multiple implementations of the network layer made it much more
difficult to change some of the code.