Simplify the way blocks are fetched remotely.
In the existing code, remote shuffle blocks are fetched as follows:
(1) An open message is sent to the remote host, which returns a
unique identifier for the set of shuffle blocks to be fetched.
(2) The shuffle blocks are requested, one at a time, from the
remote host.

With monotasks, we want more fine-grained control over how shuffle
blocks are fetched, and in particular, over the precise order
in which blocks are fetched. The open message is destructive to
this goal: if first an open message is sent for A and then for
B, it's possible when the open response comes back, the thread to
process B will be scheduled first, so B's actual block(s) will be
requested before those for A.

This commit simplifies the networking code by removing the open
message. Instead, there is a single RPC to fetch a remote block:
the host that wants a remote block sents the remote host a
BlockFetchRequest, and the remote host responds with a BlockFetchSucess
message that contains the remote block (or a BlockFetchFailure message,
if the remote block could not be found).

With this change, each block corresponds to exactly one RPC to the remote
machine. This is roughly equivalent to what happened before, because with
the old networking code, one ChunkFetchRequest was sent for each block. The
main difference with this code is that each block is treated independently,
rather than treating a set of blocks that need to be fetched as a stream of
blocks that are handled together.