Remove StorageLevel parameter when unnecessary.
This commit removes the StorageLevel parameter from DiskWriteMonotask and
from BlockManager.updateBlockInfoOnWrite, and instead automatically sets the
storage level when blocks are written to disk. This fixes an issue with the
previous code where if a block was written to memory and then written to disk,
the storage level didn't get updated properly to include disk (the old
StorageLevel parameter got ignored when the block was already stored in the
block manager).  This approach is also cleaner, because callers don't need to
explicitly specify the StorageLevel; the BlockManager can infer it.