Limit the number of concurrent macrotasks per worker.
This commit limits the number of concurrent macrotasks on each worker at
(# compute monotask slots) + (# disk monotask slots) + (# network monotask slots).
The intuition behind this is that the above value is the maximum number
of monotasks that can be run concurrently, so scheduling more macrotasks than
that will result in queueing. Queueing hurts performance because it may lead to
more memory contention, if an unnecessarily large number of macrotasks are
partially finished, and also because it can lead to load imbalance between workers,
if one worker is assigned a large number of macrotasks that end up waiting in
queues will other workers had free slots.