Write shuffle data into a DirectByteBuffer.
This commit changes the behavior of the shuffle such that the
ComputeMonotask that generates the serialized shuffle data writes all of the
data into a single DirectByteBuffer, rather than into one (non-direct) ByteBuffer
for each shuffe block. We also write an index at the beginning of the
buffer, which can be used to determine where individual shuffle blocks lie
within the file.

This change was necessary to eliminate CPU time that was occuring as part of
DiskWriteMonotasks. It took hundreds of milliseconds (for ~100MB blocks) just to
copy the data to be written from multiple buffers in the JVM to an off-heap buffer
that could be written to disk, which meant that DiskWriteMonotasks weren't fully
utilizing the disk (there were gaps at the beginning of each DiskWriteMonotask,
where the disk wasn't used at all).

This change also uses a new type of BlockId, MultipleShuffleBlocksId, to write a block
that consists of multiple shuffle blocks. This is to make it easier to differentiate
(both in coding and in debugging) between a single shuffle block and data that contains
multiple shuffle blocks.

This commit doesn't implement one thing that will be important to do eventually: it
doesn't make any attempt to try to keep track of the memory that we allocate off-heap.
We should definitely manage this somehow, so that we can avoid using too much.