Simplified parsing scripts and added drizzle functionality.
This commit removes lots of unnecessary code in the parsing scripts (e.g.,
code to attribute stragglers to particular causes) and adds functionality
to print out the speedup (considering barriers) had the job not had any
scheduler delay, time to fetch map outputs, or broadcast time, and also
the speedup if all 3 delays were eliminated.

I haven't tested this code on a real log file yet so it's likely that it
will throw a runtime error.