Add instrumentation for fetch requests.
This commit adds instrumentation for time spent fetching shuffle blocks
from remote machines. Specifically, it adds instrumentation for the remote
disk time (and then adds this time to the disk time recorded for the macrotask)
and also adds instrumentation for the total elapsed time on the remote machine
(to try to debug issues where the network isn't well utilized).