Eliminated FutureTask class.
The FutureTask class has been deleted; instead, Task has a isFutureTask
parameter, and the getFirstShuffleDep method is now in Task. FutureTask
was confusing as a class because it wrapped Task, leading to some annoying
issues where properties could be set on FutureTask but not on the inner
task (and vice versa).