Add metric for total time on executor.
This commit adds a new metric, Total Time On Executor, that describes
the total time from when the task arrived at the executor to the time
the task completion message was sent to the scheduler. In theory, this
metric is implied by the sum of the executor run time plus the various
deserialization and queueing delays, but having this variable explicitly
helps us differentiate scheduler delay from missing metrics.