Batch task launching, and add async option.
This commit does two things:

(1) Batch task launching, so if many tasks need to be launched on the same executor, all of
the tasks are sent in a single message (this is not feature-flagged).

(2) Make task launch asynchronous, so it doesn't happen in the main scheduler event
loop. This part was mostly copied from https://github.com/kayousterhout/spark/commit/1e5bced5054e0801d6285aad7cc66a9b425d95cd
but with the bit to make it thread safe removed, because as long as our executors aren't changing,
I think we're safe without it.

There is no evidence that the async task launch actually helps, so it is off by default
(and can be controlled using spark.scheduler.asyncTaskLaunch).