Account for future tasks in TaskSchedulerImpl.
This commit changes TaskSchedulerImpl to account for future tasks in
determining how many resources are left on a machine. As a result, it can
assign more future tasks to each machine each time resourceOffer()
is called (it can assign all of the outstanding future tasks in a single
call of resourceOffer(), rather than only assigning # free cores tasks
in each call).

This change has no noticable impact on the amount of scheduler delay,
but should allow us to do more batching.