Batch requests for shuffle data.
This commit changes the way that shuffle data is requested so that each each macrotask
sends one request per remote executor rather than one request per shuffle block. This commit
benefits performance because it eliminates harmful re-ordering that was occuring when shuffle
blocks were requested.  For the best performance, when executor A is sending shuffle data to
executor B, data for different tasks running on B should not be intermixed; instead, A should
send all of the data for a particular macrotask before beginning to send data for the next
macrotask. This commit guarantees that behavior in the case that all of the shuffle data is
in-memory: if executor A receives requests for data for two macrotasks, it will
receive all of the requests for blocks for one macrotask first (because all of the blocks are
described in a single request message) and the TransportRequestHandler will enqueue all of
the NetworkMonotasks for the response (by calling BlockManager.getBlockData) before handling
any responses for the other macrotask.

This differs from the current behavior: currently, we send one request for each shuffle block,
so executor A might receive requests for blocks for task 1 intermixed with requests for task 2,
so the responses would also be intermixed.