Share each machine evenly across all task sets.
With this change, the TaskSchedulerImpl tries to assign tasks to
machines such that an equal number of tasks from each outstanding
task set will be running on each machine. The motivation for this
is to try to balance the resource use on each machine so that if,
for example, a disk-intensive and CPU-intensive job are running
concurrently, a given machine will have tasks from both jobs so
that its resources remain fully utilized.  This is a heuristic to
try to get good resource usage on all machines without adding
a complex scheduling mechanism.`