Smarter queueing for disk monotasks.
This commit implements a new deficit round-robin queue for disk monotasks.
The weight of each monotask corresponds to the number of monotasks in that
particular phase of execution; e.g., if shuffle data is read by 100 different
monotasks, then each monotask will have weight 1/100.  For writing output data,
when there's just one monotask, each will have weight 1. This is so that the
resulting scheduling will end up effectively doing round-robin over different
phases of execution.

This commit also changes the disk scheduler to be altruistic, so that
monotasks for remote machines are run before monotasks for the local machine.
This is an attempt to handle slow machines.