Added disk queue that round-robins over workers.
The new disk queue has two levels of fair queueing: first, it does
round-robin over remote machines, and then within each machine,
it round-robins over monotask type. The reason for this is there
were problematic load balancing issues where one machine would
get stuck processing all of its own disk monotasks, and a result,
would block all of the remote requests for data.

As part of doing this, TaskContextImpl was changed to include the
name of the remote machine where the macrotask is running, so that
the DiskScheduler can use this information.