Increase number of concurrent tasks on each machine.
This commit changes the number of concurrent tasks that are allowed to
run on each machine to be:

2 * c_max + sum_{all resources i except max} {c_i} - 1

Where c_i is the concurrency of resource i and c_max is the max concurrency
of any resource.

This concurrency is what was proven to be sufficient in the most recent
round of proofs.